# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "一个简单的机器管理系统",
        "title": "rabbit",
        "contact": {},
        "license": {}
    },
    "paths": {
...
        "/api/v1/chart/bar": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "按照tag统计tag下的CPU个数",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/chart.APIGetChartBarOutputs"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/api/v1/chart/container/stat": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "统计平台收纳的容器使用量/cpu数/已分配机器数/已分配cpu数",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/chart.APIGetChartStatOutputs"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/api/v1/chart/pie": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "按照tag统计tag下的CPU个数",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/chart.APIGetChartBarOutputs"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/api/v1/chart/vm/stat": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "统计平台收纳的所有机器数/cpu数/已分配机器数/已分配cpu数",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/chart.APIGetChartStatOutputs"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
...
        "/api/v1/host_apply_request/assign": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "summary": "创建机器申请单处理",
                "parameters": [
                    {
                        "description": "创建机器申请单处理",
                        "name": "APIPostHostApplyRequestAssignInputs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/node.APIPostHostApplyRequestAssignInputs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/api/v1/host_apply_request/create": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "创建机器申请单",
                "parameters": [
                    {
                        "description": "创建机器申请单",
                        "name": "APIPostHostApplyRequestCreateInputs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/node.APIPostHostApplyRequestCreateInputs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/api/v1/host_apply_request/info": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "根据请求单ID获取详细信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "根据请求单ID获取详细信息",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/node.Host"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/api/v1/host_apply_request/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "机器资源申请列表",
                "parameters": [
                    {
                        "type": "string",
                        "name": "applier",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "orderBy",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/node.APIGetHostApplyRequestListOutputs"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "chart.APIGetChartBarLegend": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "chart.APIGetChartBarOutputs": {
            "type": "object",
            "properties": {
                "legend": {
                    "type": "object",
                    "$ref": "#/definitions/chart.APIGetChartBarLegend"
                },
                "series": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/chart.APIGetChartBarSeries"
                    }
                },
                "xAxis": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/chart.APIGetChartBarXAxis"
                    }
                }
            }
        },
        "chart.APIGetChartBarSeries": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "chart.APIGetChartBarXAxis": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "chart.APIGetChartStat": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "totalCpuCount": {
                    "type": "integer"
                },
                "totalHostCount": {
                    "type": "integer"
                },
                "unUsedCpuCount": {
                    "type": "integer"
                },
                "unUsedHostCount": {
                    "type": "integer"
                },
                "usedCpuCount": {
                    "type": "integer"
                },
                "usedHostCount": {
                    "type": "integer"
                }
            }
        },
        "chart.APIGetChartStatOutputs": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "subStat": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/chart.APIGetChartStat"
                    }
                },
                "totalCpuCount": {
                    "type": "integer"
                },
                "totalHostCount": {
                    "type": "integer"
                },
                "unUsedCpuCount": {
                    "type": "integer"
                },
                "unUsedHostCount": {
                    "type": "integer"
                },
                "usedCpuCount": {
                    "type": "integer"
                },
                "usedHostCount": {
                    "type": "integer"
                }
            }
        },
...
        "node.APIGetHostApplyRequestListOutputs": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/node.APIGetHostApplyRequestListData"
                    }
                },
                "totalCount": {
                    "type": "integer"
                }
            }
        },
...
        "node.APIPostHostApplyRequestAssignInputs": {
            "type": "object",
            "properties": {
                "hostIDs": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "state": {
                    "type": "string"
                },
                "tagIDs": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "node.APIPostHostApplyRequestCreateInputs": {
            "type": "object",
            "properties": {
                "applier": {
                    "type": "string"
                },
                "cpu": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "memory": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "releaseAt": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                }
            }
        },
...
        "node.Host": {
            "type": "object",
            "properties": {
                "applyUser": {
                    "type": "string"
                },
                "areaName": {
                    "type": "string"
                },
                "cloudPoolCode": {
                    "type": "string"
                },
                "cloudPoolName": {
                    "type": "string"
                },
                "coreTotalNum": {
                    "type": "integer"
                },
                "cpuNumber": {
                    "type": "integer"
                },
                "cpuUsage": {
                    "type": "number"
                },
                "createTime": {
                    "type": "string"
                },
                "databaseVersion": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "deployDate": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "devAreaCode": {
                    "type": "string"
                },
                "devCenterName": {
                    "type": "string"
                },
                "devOwner": {
                    "type": "string"
                },
                "devTypeCode": {
                    "type": "string"
                },
                "fsUsage": {
                    "type": "number"
                },
                "funDesc": {
                    "type": "string"
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/node.HostGroup"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "instanceId": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "isWarning": {
                    "type": "boolean"
                },
                "logicSystem": {
                    "type": "string"
                },
                "logicSystemCnName": {
                    "type": "string"
                },
                "manIp": {
                    "type": "string"
                },
                "managerA": {
                    "type": "string"
                },
                "managerB": {
                    "type": "string"
                },
                "memorySize": {
                    "type": "integer"
                },
                "memoryUsage": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "oracle": {
                    "type": "string"
                },
                "osVersion": {
                    "type": "string"
                },
                "partType": {
                    "type": "string"
                },
                "partTypeCode": {
                    "type": "string"
                },
                "physicalSystem": {
                    "type": "string"
                },
                "physicalSystemArea": {
                    "type": "string"
                },
                "physicalSystemCnName": {
                    "type": "string"
                },
                "physicalSystemEnName": {
                    "type": "string"
                },
                "servSpaceCodeList": {
                    "type": "string"
                },
                "servSpaceNameList": {
                    "type": "string"
                },
                "setupMode": {
                    "type": "string"
                },
                "srvStatus": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app.Tag"
                    }
                },
                "type": {
                    "type": "string"
                },
                "updateTime": {
                    "type": "string"
                },
                "virtFcNum": {
                    "type": "string"
                },
                "virtNetNum": {
                    "type": "string"
                }
            }
        },
        "node.HostGroup": {
            "type": "object",
            "properties": {
                "caasServiceId": {
                    "type": "integer"
                },
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/node.HostGroup"
                    }
                },
                "childrenHostCount": {
                    "type": "integer"
                },
                "childrenPodCount": {
                    "type": "integer"
                },
                "createUser": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isWarning": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "parentID": {
                    "type": "integer"
                },
                "parentName": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "pathArray": {
                    "type": "string"
                },
                "relatedHostCount": {
                    "type": "integer"
                },
                "relatedPodCount": {
                    "type": "integer"
                },
                "subGroupCount": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    }
}